+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -4,6 +4,33 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v3.2.0
+------------------------------
+*October 25, 2021*
+
+### Changed
+- Button border style definition moved to button types. Introduced `background-color()` mixin to apply the same colour for the background and for the border colour.
+- Each button size vertical padding reduced by 1px to make button height 56px/48px/40px/32px instead of 58px/50px/42px/34px.
+
+### Removed
+- Outline css styles as focus styles applied with `box-shadow`.
+
+
+v3.1.0
+------------------------------
+*October 22, 2021*
+
+### Changed
+- Add spacing to buttons following a paragraph tag. This matches the current behaviour in Fozzie.
+- Removed the background from the link button in disabled state.
+- `f-vue-icons` package version bump.
+- Button story now shows all the types and sizes of the button component (primary, secondary, outline and ghost).
+- Changed knobs to controls for switching button props.
+- Separate Demo story to still show one button in isolation.
+- Separate story for Link Button to be able to run tests properly.
+- Updated the tests in line with the changes.
+
+
 v3.0.2
 ------------------------------
 *October 13, 2021*